0.0.5 fix: в классе Event исправлены типы полей eventId, userId, entityId с long на int

+++ src/main/java/ru/yandex/practicum/filmorate/model/Event.java
@@ -18,11 +18,11 @@ public class Event {
     @JsonFormat(shape = JsonFormat.Shape.NUMBER)
     Date timestamp;
     @NotNull
-    long userId;
+    int userId;
     String eventType;
     String operation;
     @NotNull
-    long eventId;
+    int eventId;
     @NotNull
-    long entityId;
+    int entityId;
 }
